Restructure packages into 3 high-level groups: collector, runner, scripts.
