+++ datarepo-clienttests/src/main/java/common/BasicStatistics.java
@@ -1,0 +1,4 @@
+package common;
+
+public class BasicStatistics {
+}
