+++ datarepo-clienttests/src/main/java/scripts/disruptivescripts/RandomPodDelete.java
@@ -1,4 +1,4 @@
-package disruptivescripts;
+package scripts.disruptivescripts;
 
 import java.util.List;
 import java.util.concurrent.TimeUnit;
