+++ datarepo-clienttests/src/main/java/scripts/testscripts/DatasetCustodianPermissions.java
@@ -1,4 +1,4 @@
-package testscripts;
+package scripts.testscripts;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 
@@ -17,7 +17,7 @@ import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import runner.config.TestUserSpecification;
-import testscripts.baseclasses.SimpleDataset;
+import scripts.testscripts.baseclasses.SimpleDataset;
 import utils.BigQueryUtils;
 import utils.DataRepoUtils;
 import utils.SAMUtils;
