+++ datarepo-clienttests/src/main/java/scripts/testscripts/IngestFile.java
@@ -1,4 +1,4 @@
-package testscripts;
+package scripts.testscripts;
 
 import bio.terra.datarepo.api.RepositoryApi;
 import bio.terra.datarepo.client.ApiClient;
@@ -13,7 +13,7 @@ import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import runner.config.TestUserSpecification;
-import testscripts.baseclasses.SimpleDataset;
+import scripts.testscripts.baseclasses.SimpleDataset;
 import utils.DataRepoUtils;
 import utils.FileUtils;
