+++ datarepo-clienttests/src/main/java/scripts/testscripts/RetrieveSnapshot.java
@@ -1,45 +1,35 @@
-package testscripts;
+package scripts.testscripts;
 
 import bio.terra.datarepo.api.RepositoryApi;
 import bio.terra.datarepo.client.ApiClient;
 import bio.terra.datarepo.model.BulkLoadArrayRequestModel;
 import bio.terra.datarepo.model.BulkLoadArrayResultModel;
 import bio.terra.datarepo.model.BulkLoadFileModel;
-import bio.terra.datarepo.model.DataDeletionGcsFileModel;
-import bio.terra.datarepo.model.DataDeletionRequest;
-import bio.terra.datarepo.model.DataDeletionTableModel;
-import bio.terra.datarepo.model.DatasetModel;
 import bio.terra.datarepo.model.DeleteResponseModel;
 import bio.terra.datarepo.model.IngestRequestModel;
 import bio.terra.datarepo.model.IngestResponseModel;
 import bio.terra.datarepo.model.JobModel;
-import com.google.api.client.util.Charsets;
-import com.google.cloud.bigquery.BigQuery;
-import com.google.cloud.bigquery.TableResult;
-import java.io.ByteArrayOutputStream;
+import bio.terra.datarepo.model.SnapshotModel;
+import bio.terra.datarepo.model.SnapshotSummaryModel;
 import java.net.URI;
 import java.nio.charset.StandardCharsets;
-import java.util.Collections;
 import java.util.List;
-import java.util.stream.Collectors;
-import java.util.stream.StreamSupport;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import runner.config.TestUserSpecification;
-import testscripts.baseclasses.SimpleDataset;
-import utils.BigQueryUtils;
+import scripts.testscripts.baseclasses.SimpleDataset;
 import utils.DataRepoUtils;
 import utils.FileUtils;
 
-public class SoftDeleteDataset extends SimpleDataset {
-  private static final Logger logger = LoggerFactory.getLogger(SoftDeleteDataset.class);
+public class RetrieveSnapshot extends SimpleDataset {
+  private static final Logger logger = LoggerFactory.getLogger(RetrieveSnapshot.class);
 
   /** Public constructor so that this class can be instantiated via reflection. */
-  public SoftDeleteDataset() {
+  public RetrieveSnapshot() {
     super();
   }
 
-  private DataDeletionRequest dataDeletionRequest;
+  private SnapshotSummaryModel snapshotSummaryModel;
 
   private String testConfigGetIngestbucket;
 
@@ -47,7 +37,7 @@ public class SoftDeleteDataset extends SimpleDataset {
     // create the profile and dataset
     super.setup(testUsers);
 
-    // get the ApiClient for the dataset creator
+    // get the ApiClient for the snapshot creator, same as the dataset creator
     ApiClient datasetCreatorClient = DataRepoUtils.getClientForTestUser(datasetCreator, server);
     RepositoryApi repositoryApi = new RepositoryApi(datasetCreatorClient);
 
@@ -58,15 +48,15 @@ public class SoftDeleteDataset extends SimpleDataset {
     // ingest a file -- TODO CannedTestData.getMeA1KBFile
     URI sourceUri = new URI("gs://jade-testdata/fileloadprofiletest/1KBfile.txt");
 
-    String targetPath = "/testrunner/softDel/" + FileUtils.randomizeName("") + ".txt";
+    String targetPath = "/testrunner/IngestFile/" + FileUtils.randomizeName("") + ".txt";
 
     BulkLoadFileModel fileLoadModel =
         new BulkLoadFileModel()
             .sourcePath(sourceUri.toString())
-            .description("softDel")
+            .description("IngestFile")
             .mimeType("text/plain")
             .targetPath(targetPath);
-    String loadTag = FileUtils.randomizeName("softDelTest");
+    String loadTag = FileUtils.randomizeName("lookupTest");
     BulkLoadArrayRequestModel fileLoadModelArray =
         new BulkLoadArrayRequestModel()
             .profileId(datasetSummaryModel.getDefaultProfileId())
@@ -90,7 +80,7 @@ public class SoftDeleteDataset extends SimpleDataset {
             + "\"}\n";
     byte[] fileRefBytes = jsonLine.getBytes(StandardCharsets.UTF_8);
     String jsonFileName = FileUtils.randomizeName("this-better-pass") + ".json";
-    String dirInCloud = "scratch/softDel";
+    String dirInCloud = "scratch/testRetrieveSnapshot/";
     String fileRefName = dirInCloud + "/" + jsonFileName;
     String gsPath = FileUtils.createGsPath(fileRefBytes, fileRefName, testConfigGetIngestbucket);
 
@@ -109,74 +99,42 @@ public class SoftDeleteDataset extends SimpleDataset {
             repositoryApi, ingestTabularDataJobResponse, IngestResponseModel.class);
     logger.info("Successfully loaded data into dataset: {}", ingestResponse.getDataset());
 
-    String datasetId = datasetSummaryModel.getId();
-    DatasetModel datasetModel = repositoryApi.retrieveDataset(datasetId);
-    String dataProject = datasetModel.getDataProject();
-    String tableName = datasetModel.getSchema().getTables().get(0).getName();
+    // make the create snapshot request and wait for the job to finish
+    JobModel createSnapshotJobResponse =
+        DataRepoUtils.createSnapshot(
+            repositoryApi, datasetSummaryModel, "snapshot-simple.json", true);
 
-    // get row ids for table
-    String sqlQuery =
-        BigQueryUtils.buildSelectQuery(
-            dataProject,
-            BigQueryUtils.getDatasetName(datasetModel.getName()),
-            tableName,
-            "datarepo_row_id",
-            1L);
-    BigQuery bigQueryClient =
-        BigQueryUtils.getClientForTestUser(datasetCreator, datasetModel.getDataProject());
-    TableResult result = BigQueryUtils.queryBigQuery(bigQueryClient, sqlQuery);
-    List<String> rowIds =
-        StreamSupport.stream(result.getValues().spliterator(), false)
-            .map(fieldValues -> fieldValues.get(0).getStringValue())
-            .collect(Collectors.toList());
-
-    logger.info("Successfully retrieved row ids for table: {}", tableName);
-
-    // write to GCS
-    String csvFileName = FileUtils.randomizeName("this-too-better-pass") + ".csv";
-    String csvFileRefName = dirInCloud + "/" + csvFileName;
-
-    ByteArrayOutputStream output = new ByteArrayOutputStream();
-    for (String line : rowIds) {
-      output.write((line + "\n").getBytes(Charsets.UTF_8));
-    }
-    byte[] bytes = output.toByteArray();
-    String gcsPath = FileUtils.createGsPath(bytes, csvFileRefName, testConfigGetIngestbucket);
-
-    // build the deletion request with pointers to the file with row ids to soft delete
-    DataDeletionGcsFileModel deletionGcsFileModel =
-        new DataDeletionGcsFileModel()
-            .fileType(DataDeletionGcsFileModel.FileTypeEnum.CSV)
-            .path(gcsPath);
-    DataDeletionTableModel deletionTableFile =
-        new DataDeletionTableModel().tableName(tableName).gcsFileSpec(deletionGcsFileModel);
-    List<DataDeletionTableModel> dataDeletionTableModels =
-        Collections.singletonList(deletionTableFile);
-    dataDeletionRequest =
-        new DataDeletionRequest()
-            .deleteType(DataDeletionRequest.DeleteTypeEnum.SOFT)
-            .specType(DataDeletionRequest.SpecTypeEnum.GCSFILE)
-            .tables(dataDeletionTableModels);
+    // save a reference to the snapshot summary model so we can delete it in cleanup()
+    snapshotSummaryModel =
+        DataRepoUtils.expectJobSuccess(
+            repositoryApi, createSnapshotJobResponse, SnapshotSummaryModel.class);
+    logger.info("Successfully created snapshot: {}", snapshotSummaryModel.getName());
   }
 
   public void userJourney(TestUserSpecification testUser) throws Exception {
     ApiClient apiClient = DataRepoUtils.getClientForTestUser(testUser, server);
     RepositoryApi repositoryApi = new RepositoryApi(apiClient);
 
-    // send off the soft delete request
-    JobModel softDeleteJobResponse =
-        repositoryApi.applyDatasetDataDeletion(datasetSummaryModel.getId(), dataDeletionRequest);
-    softDeleteJobResponse = DataRepoUtils.waitForJobToFinish(repositoryApi, softDeleteJobResponse);
-    DeleteResponseModel deleteResponseModel =
-        DataRepoUtils.expectJobSuccess(
-            repositoryApi, softDeleteJobResponse, DeleteResponseModel.class);
+    SnapshotModel snapshotModel = repositoryApi.retrieveSnapshot(snapshotSummaryModel.getId());
     logger.debug(
-        "Successfully soft deleted rows from dataset: {} with state {}",
-        datasetSummaryModel.getName(),
-        deleteResponseModel.getObjectState());
+        "Successfully retrieved snaphot: {}, data project: {}",
+        snapshotModel.getName(),
+        snapshotModel.getDataProject());
   }
 
   public void cleanup(List<TestUserSpecification> testUsers) throws Exception {
+    // get the ApiClient for the dataset creator
+    ApiClient datasetCreatorClient = DataRepoUtils.getClientForTestUser(datasetCreator, server);
+    RepositoryApi repositoryApi = new RepositoryApi(datasetCreatorClient);
+
+    // make the delete request and wait for the job to finish
+    JobModel deleteSnapshotJobResponse = repositoryApi.deleteSnapshot(snapshotSummaryModel.getId());
+    deleteSnapshotJobResponse =
+        DataRepoUtils.waitForJobToFinish(repositoryApi, deleteSnapshotJobResponse);
+    DataRepoUtils.expectJobSuccess(
+        repositoryApi, deleteSnapshotJobResponse, DeleteResponseModel.class);
+    logger.info("Successfully deleted snapshot: {}", snapshotSummaryModel.getName());
+
     // delete the profile and dataset
     super.cleanup(testUsers);
