+++ datarepo-clienttests/src/main/java/scripts/testscripts/ScalePodsToZero.java
@@ -1,29 +1,29 @@
-package testscripts;
+package scripts.testscripts;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 
 import bio.terra.datarepo.api.RepositoryApi;
 import bio.terra.datarepo.client.ApiClient;
-import bio.terra.datarepo.model.BulkLoadArrayRequestModel;
-import bio.terra.datarepo.model.BulkLoadResultModel;
-import bio.terra.datarepo.model.JobModel;
+import bio.terra.datarepo.client.ApiException;
+import bio.terra.datarepo.model.*;
 import java.util.List;
+import java.util.concurrent.TimeUnit;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import runner.config.TestUserSpecification;
-import testscripts.baseclasses.SimpleDataset;
+import scripts.testscripts.baseclasses.SimpleDataset;
 import utils.BulkLoadUtils;
 import utils.DataRepoUtils;
 import utils.KubernetesClientUtils;
 
-public class PodDelete extends SimpleDataset {
+public class ScalePodsToZero extends SimpleDataset {
   private static final Logger logger = LoggerFactory.getLogger(ScalePodsToZero.class);
 
   /** Public constructor so that this class can be instantiated via reflection. */
-  public PodDelete() {
+  public ScalePodsToZero() {
     super();
-    manipulatesKubernetes = true; // this test script manipulates Kubernetess
+    manipulatesKubernetes = true; // this test script manipulates Kubernetes
   }
 
   private int filesToLoad;
@@ -38,7 +38,7 @@ public class PodDelete extends SimpleDataset {
   }
 
   // The purpose of this test is to have a long-running workload that completes successfully
-  // while we delete a random pod.
+  // while we scale pods to zero and then scale them back up.
   public void userJourney(TestUserSpecification testUser) throws Exception {
     ApiClient apiClient = DataRepoUtils.getClientForTestUser(testUser, server);
     RepositoryApi repositoryApi = new RepositoryApi(apiClient);
@@ -58,9 +58,41 @@ public class PodDelete extends SimpleDataset {
         DataRepoUtils.pollForRunningJob(repositoryApi, bulkLoadArrayJobResponse, 30);
 
     if (bulkLoadArrayJobResponse.getJobStatus().equals(JobModel.JobStatusEnum.RUNNING)) {
-      KubernetesClientUtils.deleteRandomPod();
-    } else {
-      throw new Exception("Job finished before we were able to test the delete functionality.");
+      logger.debug("Scaling pods down to 0");
+      KubernetesClientUtils.changeReplicaSetSizeAndWait(0);
+
+      try {
+        bulkLoadArrayJobResponse =
+            DataRepoUtils.pollForRunningJob(repositoryApi, bulkLoadArrayJobResponse, 30);
+      } catch (ApiException ex) {
+        logger.debug(
+            "Catching expected exception while pod size = 0, Job Status: {}",
+            bulkLoadArrayJobResponse.getJobStatus());
+      }
+
+      logger.debug("Scaling pods back up to 3.");
+      KubernetesClientUtils.changeReplicaSetSizeAndWait(3);
+      int retryCounter = 0;
+      // give the job a few chances to get a non-failing results while the pods are scaled back up.
+      ApiException lastException = null;
+      while (bulkLoadArrayJobResponse.getJobStatus().equals(JobModel.JobStatusEnum.RUNNING)
+          && retryCounter < 10) {
+        retryCounter++;
+        try {
+          bulkLoadArrayJobResponse =
+              DataRepoUtils.pollForRunningJob(repositoryApi, bulkLoadArrayJobResponse, 30);
+          lastException = null;
+        } catch (ApiException ex) {
+          logger.debug(
+              "Catching expected error while we wait for pods to scale back up. Retry # {}",
+              retryCounter);
+          lastException = ex;
+          TimeUnit.SECONDS.sleep(30);
+        }
+      }
+      if (lastException != null) {
+        throw lastException;
+      }
     }
     // =========================================================================
